씬 클래스 구현 및 씬을 로드하여 적용하도록 수정

+++ GameEngine_SOURCE/geGameObject2.cpp
@@ -21,7 +21,7 @@ namespace ge
 
 	void GameObject2::Update()
 	{
-		const int speed = 150.0f;
+		const float speed = 150.0f;
 		if (mGubun == 1)
 		{
 			if (Input::GetKey(eKeyCode::Left))
